Account settings in progress, rebuilding with user id's instead and fixed unlink bug with profile image when none set.

+++ backend/account/profile-img.php
@@ -6,10 +6,11 @@ if (!isset($_SESSION)) {
  include $_SERVER["DOCUMENT_ROOT"] . "/system/connection.php";
 
  $email = $_SESSION['loggedin'];
+ $user_id = $_SESSION['user_id'];
 
 
-     $stmt_edit = $pdo->prepare('SELECT userPic FROM users WHERE email=:email');
-     $stmt_edit->execute(array(':email'=>$email));
+     $stmt_edit = $pdo->prepare('SELECT userPic FROM users WHERE userid=:userid');
+     $stmt_edit->execute(array(':userid'=>$user_id));
      $edit_row = $stmt_edit->fetch(PDO::FETCH_ASSOC);
      extract($edit_row);
 
@@ -22,10 +23,12 @@ if (!isset($_SESSION)) {
          $upload_dir = '../../frontend/user_images/'; // upload directory
    $imgExt = strtolower(pathinfo($imgFile, PATHINFO_EXTENSION)); // get image extension
    $valid_extensions = array('jpeg', 'jpg', 'png', 'gif'); // valid extensions
-   $userpic = $email.".".$imgExt;
+   $userpic = $user_id.".".$imgExt;
          if (in_array($imgExt, $valid_extensions)) {
              if ($imgSize < 5000000) {
-                 unlink($upload_dir.$edit_row['userPic']);
+                 if (!empty($edit_row['userPic'])) {
+                     unlink($upload_dir.$edit_row['userPic']);
+                 }
                  move_uploaded_file($tmp_dir, $upload_dir.$userpic);
              } else {
                  $errMSG = "Sorry, your file is too large it should be less then 5MB";
@@ -43,16 +46,12 @@ if (!isset($_SESSION)) {
      if (!isset($errMSG)) {
          $stmt = $pdo->prepare('UPDATE users
               SET userPic=:upic
-               WHERE email=:email');
-         $stmt->bindParam(':email', $email);
+               WHERE userid=:userid');
+         $stmt->bindParam(':userid', $user_id);
          $stmt->bindParam(':upic', $userpic);
          if ($stmt->execute()) {
-             ?>
-                <script>
-    alert('Successfully Updated ...');
-    window.location.href='../../profile-settings';
-    </script>
-                <?php
+             $_SESSION['Image_Updated'] = "Successfully Updated ...";
+             header("location: ../../profile-settings");
          } else {
              $errMSG = "Sorry Data Could Not Updated !";
          }
